+++ CHANGELOG.md
@@ -6,14 +6,17 @@ This file is used to track unreleased changes, for a complete history, check the
 
 ### ✨ Added
 
+* There is new consolidated secret generator
+
 ### 🏗 Changed
 
 * Upgrade Kustomize API to 0.5.1
+* Upgrade Berglas to 0.5.3
 
 ### ⏳ Deprecated
 
-* Berglas support is deprecated and will be removed in the next release
-* Random secret generation is deprecated and will be removed in the next release
+* Berglas support is deprecated (in favor of Secret Manager) and will be removed in the next release
+* Random secret generation is deprecated (in favor the new Secret generator) and will be removed in the next release
 
 ### 🛑 Removed